refactor(venture): extract neutral Cairo bridge
