+++ .github/workflows/ci.yml
@@ -823,7 +823,7 @@ jobs:
             throw "Windows App Runtime installer failed with exit code $LASTEXITCODE"
           }
           Set-Location code/packages/rust
-          cargo test -p paint-vm-cairo -p venture-browser-qt -p venture-browser-windows
+          cargo test -p paint-vm-cairo -p venture-browser-cairo -p venture-browser-qt -p venture-browser-windows
           if ($LASTEXITCODE -ne 0) {
             throw "Venture Windows render and bridge tests failed with exit code $LASTEXITCODE"
           }
