+++ code/packages/rust/Cargo.toml
@@ -170,6 +170,7 @@ members = [
     "html-to-layout",
     "html-to-paint",
     "venture-browser-core",
+    "venture-browser-cairo",
     "venture-browser-macos",
     "venture-browser-qt",
     "venture-browser-windows",
