+++ code/packages/rust/venture-browser-cairo/BUILD
@@ -1,0 +1,2 @@
+cargo test --manifest-path Cargo.toml
+cargo build --manifest-path Cargo.toml
