+++ code/packages/rust/venture-browser-cairo/CHANGELOG.md
@@ -1,0 +1,6 @@
+# Changelog
+
+## Unreleased
+
+- Extract the shared Venture browser controller, Cairo renderer, and native C
+  ABI compatibility exports from the Qt adapter into a backend-neutral crate.
