+++ code/packages/rust/venture-browser-qt/Cargo.toml
@@ -2,18 +2,13 @@
 name = "venture-browser-qt"
 version = "0.1.0"
 edition = "2021"
-description = "Cairo content bridge for Venture's Mosaic-generated Qt Quick host."
+description = "Thin Qt adapter facade for Venture's backend-neutral Cairo content bridge."
 
 [lib]
-crate-type = ["rlib", "cdylib"]
+crate-type = ["rlib"]
 
 [dependencies]
-html-to-layout = { path = "../html-to-layout" }
-html-to-paint = { path = "../html-to-paint" }
-layout-text-measure-native = { path = "../layout-text-measure-native" }
-paint-vm-cairo = { path = "../paint-vm-cairo" }
-text-native = { path = "../text-native" }
-venture-browser-core = { path = "../venture-browser-core" }
+venture-browser-cairo = { path = "../venture-browser-cairo" }
 
 [dev-dependencies]
 mosaic-package-artifact-builder = { path = "../mosaic-package-artifact-builder" }
