+++ code/programs/mosaic/venture-browser/tests/package_compiles.rs
@@ -518,6 +518,8 @@ fn backend_build_scripts_cover_the_complete_matrix_and_direct_builds() {
         "swift build",
         "cargo \"${bridge_args[@]}\"",
         "libventure_browser_macos.dylib",
+        "venture-browser-cairo",
+        "libventure_browser_cairo",
         "venture-browser-qt",
         "libventure_browser_qt",
         "qt_project_launch",
@@ -541,6 +543,8 @@ fn backend_build_scripts_cover_the_complete_matrix_and_direct_builds() {
         "swift",
         "venture-browser-macos",
         "libventure_browser_macos.dylib",
+        "venture-browser-cairo",
+        "venture_browser_cairo",
         "venture-browser-qt",
         "venture_browser_qt",
         "qt_project_launch",
